+++ src/main/java/me/myogoo/myotus/client/gui/MyoIcon.java
@@ -8,7 +8,10 @@ public enum MyoIcon {
 
     BUG_REPORT(0, 0),
     MYOTUS_CONFIG(16,0),
-    AE2TB_CONFIG(32,0);
+    AE2TB_CONFIG(32,0),
+    AE2FCT_CONFIG(48,0),
+    SHOW_UPGRADE_PANEL(0,16),
+    HIDE_UPGRADE_PANEL(0,32);
     private static final ResourceLocation TEXTURE = Myotus.makeId("textures/gui/myoicons.png");
 
     private final int x;