+++ src/main/java/me/myogoo/myotus/impl/MyotusAPIImpl.java
@@ -3,7 +3,9 @@ package me.myogoo.myotus.impl;
 import me.myogoo.myotus.api.IMyotusAPI;
 import me.myogoo.myotus.api.integration.IModIntegrationManager;
 import me.myogoo.myotus.api.registrar.IConfigRegistrar;
+import me.myogoo.myotus.api.registrar.ICreativeTabRegistrar;
 import me.myogoo.myotus.api.registrar.IModRegistrar;
+import me.myogoo.myotus.impl.registrar.CreativeTabRegistrarImpl;
 import me.myogoo.myotus.impl.integration.ModIntegrationManagerImpl;
 import me.myogoo.myotus.impl.registrar.ConfigRegistrarImpl;
 import me.myogoo.myotus.impl.registrar.ModRegistrarImpl;
@@ -13,6 +15,7 @@ public enum MyotusAPIImpl implements IMyotusAPI {
 
     private final IModRegistrar modRegistrar = new ModRegistrarImpl();
     private final IConfigRegistrar configRegistrar = new ConfigRegistrarImpl();
+    private final ICreativeTabRegistrar creativeTabRegistrar = new CreativeTabRegistrarImpl();
     private final IModIntegrationManager modIntegrationManager = new ModIntegrationManagerImpl();
 
     @Override
@@ -23,6 +26,11 @@ public enum MyotusAPIImpl implements IMyotusAPI {
     @Override
     public IConfigRegistrar configRegistrar() {
         return configRegistrar;
+    }
+
+    @Override
+    public ICreativeTabRegistrar creativeTabRegistrar() {
+        return creativeTabRegistrar;
     }
 
     @Override